backend/requirements.txt
Authlib==1.6.5
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
google-generativeai>=0.8.3
spacy>=3.8.0
pypdf>=3.1.0